requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-core
huggingface-hub
sentence-transformers
faiss-cpu
openai
tqdm
pypdf
PyMuPDF
pdfminer.six
pandas
numpy
typing-extensions
